Consolidate token propagation code
